+++ src/App.tsx
@@ -1,52 +1,17 @@
-import { FormEvent, useState } from 'react'
+import { Route, Routes } from 'react-router-dom'
 import './App.css'
-import Greeting from './components/Greeting'
 import Navbar from './components/Navbarheading'
-import Post from './components/Post'
-import usePosts from './hooks/usePost'
+import Home from './pages/Home'
+import Profile from './pages/Profile'
 
 function App() {
-  const { posts, isLoading, isSubmitting, createPost } = usePosts()
-  const [newTitle, setNewTitle] = useState<string>('')
-  const [newBody, setNewBody] = useState<string>('')
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault()
-
-    try {
-      await createPost(newTitle, newBody)
-
-      setNewTitle('')
-      setNewBody('')
-    } catch (err) {
-      console.error(err)
-    }
-  }
-
-  if (isLoading) return <h1>Loading...</h1>
-
   return (
     <div className="App">
       <Navbar />
-      <Greeting name="Bun" isLoggedIn={true} />
-
-      <form onSubmit={handleSubmit}>
-        <label>Title</label>
-        <input type="text" value={newTitle} onChange={(e) => setNewTitle(e.target.value)} required />
-        <label>Body</label>
-        <input type="text" value={newBody} onChange={(e) => setNewBody(e.target.value)} required />
-
-        <button type="submit" disabled={isSubmitting}>
-          {isSubmitting ? 'Submitting...' : 'Submit'}
-        </button>
-      </form>
-
-      <div className="feed-container">
-        {posts &&
-          posts.map((post) => {
-            return <Post key={post.id} post={post} />
-          })}
-      </div>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/profile" element={<Profile />} />
+      </Routes>
     </div>
   )
 }
