+++ src/pages/Profile.tsx
@@ -1,0 +1,11 @@
+import Greeting from '../components/Greeting'
+
+const Profile = () => {
+  return (
+    <div>
+      <Greeting name="Men" isLoggedIn={true} />
+    </div>
+  )
+}
+
+export default Profile
